+++ releases/ChangeLog.md
@@ -1,5 +1,6 @@
 ## 5.1.3 - 2023-06-20
 
+- New Feature: Batch Sign PSBTs. `Advanced/Tools -> File Management -> Batch Sign`
 - Enhancement: change Key Origin Information export format in multisig `addresses.csv` to match
   [BIP-0380](https://github.com/bitcoin/bips/blob/master/bip-0380.mediawiki#key-expressions)
   was `(m=0F056943)/m/48'/1'/0'/2'/0/0` now `[0F056943/48'/1'/0'/2'/0/0]`